fix: migrate OrderProgressEventsUpdater because of tests dependency
