+++ apps/cowswap-frontend/src/modules/orderProgressBar/updaters/OrderProgressEventsUpdater.test.ts
@@ -1,6 +1,6 @@
 import type { OrderFillability } from 'modules/ordersTable'
 
-import { computeUnfillableOrderIds, getNewlyFillableOrderIds } from './OrderProgressEventsUpdater'
+import { computeUnfillableOrderIds, getNewlyFillableOrderIds } from './utils'
 
 type TestOrder = {
   id: string
